xmpp: in 'fill' mode, smart calls provides in simulation.
This permits to simulate state changes in order to get the right state
when a lifecycle is used by several components. For instance, Haproxy
can be used by several components. We don't want to specify several
time the Galera cluster. This is only required the first time that
Haproxy is used.